backend/share/requirements.txt
# FastAPI - a modern, fast (high-performance), web framework for building APIs with Python 3.7+ based on standard Python type hints.
fastapi==0.108.0

# SQLAlchemy - the Python SQL toolkit and Object-Relational Mapping (ORM) library.
sqlalchemy==2.0.25

# pyodbc - a Python DB API 2 module for ODBC.
pyodbc==5.1.0

# Uvicorn - an ASGI web server implementation for Python.
uvicorn==0.25.0

# Redis - a Python client for Redis key-value store.
redis==5.0.1

# python-dotenv - a Python module that reads key-value pairs from a .env file and sets them as environment variables.
python-dotenv==1.0.0

# Loguru - a library which aims to bring enjoyable logging in Python.
loguru==0.7.2

# Pydantic - a data validation and settings management using Python type annotations.
pydantic==2.5.3

# python-jose - a JavaScript Object Signing and Encryption (JOSE) implementation in Python.
python-jose[cryptography]==3.3.0

# argon2-cffi - a Python CFFI bindings for the Argon2 password hashing algorithm.
argon2-cffi==23.1.0

# alembic - a lightweight database migration tool for usage with the SQLAlchemy Database Toolkit for Python.
alembic==1.13.1

# pytest - a testing framework for Python.
pytest==8.1.1

# requests - a Python HTTP library.
requests==2.31.0

# python-multipart - a Python module for parsing multipart/form-data.
python-multipart==0.0.9